server/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.4
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
Pillow==4.3.0
Pygments==2.2.0
PyJWT==1.5.2
PySocks==1.6.7
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
six==1.11.0
twilio==6.8.3
uritemplate==3.0.0
urllib3==1.22
